+++ CHANGELOG.md
@@ -26,6 +26,9 @@
     database when the group is deleted.
   * Fixed race conditions in the `setText`, `appendText`, and `restoreRevision`
     functions.
+  * Added an optional `authorId` parameter to `appendText`,
+    `copyPadWithoutHistory`, `createGroupPad`, `createPad`, `restoreRevision`,
+    `setHTML`, and `setText`, and bumped the latest API version to 1.3.0.
 * Fixed a crash if the database is busy enough to cause a query timeout.
 * New `/health` endpoint for getting information about Etherpad's health (see
   [draft-inadarei-api-health-check-06](https://www.ietf.org/archive/id/draft-inadarei-api-health-check-06.html)).